pyproject.toml
[project]
name = "dme-sync"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "openai (>=1.79.0,<2.0.0)",
    "typesense (>=1.0.3,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "selectolax (>=0.3.29,<0.4.0)",
    "boto3 (>=1.38.18,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "weaviate-client (>=4.14.4,<5.0.0)",
    "spacy (>=3.8.6,<4.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "pinecone (>=7.0.0,<8.0.0)",
    "neo4j (>=5.28.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.10"
psycopg2-binary = "^2.9.10"
jsonschema = "^4.23.0"
neo4j = "^5.28.1"
pre-commit = "^4.2.0"

